# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  wallet.Stock:
    properties:
      brokerId:
        type: string
      commission:
        type: number
      date:
        type: string
      id:
        type: string
      itemType:
        type: string
      portfolioId:
        type: string
      price:
        type: number
      shares:
        type: number
      symbol:
        type: string
    required:
    - brokerId
    - date
    - itemType
    - portfolioId
    - price
    - shares
    - symbol
    type: object
  wallet.StockFund:
    properties:
      brokerId:
        type: string
      commission:
        type: number
      date:
        type: string
      id:
        type: string
      itemType:
        type: string
      portfolioId:
        type: string
      price:
        type: number
      shares:
        type: number
      symbol:
        type: string
    required:
    - brokerId
    - date
    - itemType
    - portfolioId
    - price
    - shares
    - symbol
    type: object
  wallet.TreasuryDirect:
    properties:
      brokerId:
        type: string
      commission:
        type: number
      date:
        type: string
      fixedInterestRate:
        description: DueDate           *time.Time `json:"dueDate" bson:"dueDate" validate:"required"`
        type: number
      id:
        type: string
      itemType:
        type: string
      portfolioId:
        type: string
      price:
        type: number
      shares:
        type: number
      symbol:
        type: string
    required:
    - brokerId
    - date
    - fixedInterestRate
    - itemType
    - portfolioId
    - price
    - shares
    - symbol
    type: object
host: localhost:8889
info:
  contact:
    name: API Support
    url: https://github.com/mfinancecombr/finance-wallet-api
  description: mfinance Wallet API data.
  license:
    name: BSD 3-Clause
    url: https://opensource.org/licenses/BSD-3-Clause
  title: MFinance Wallet API
  version: 0.1.0
paths:
...
  /stocks-funds/sales/{id}:
    get:
      consumes:
      - application/json
      description: get stocks fund sale data
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.StockFund'
      summary: Get stocks fund sale by ID
    put:
      consumes:
      - application/json
      description: update new stocks fund sale
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some stocks fund sale
...
  /stocks/purchases/{id}:
    get:
      consumes:
      - application/json
      description: get stocks purchase data
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Stock'
      summary: Get stocks purchase by ID
    put:
      consumes:
      - application/json
      description: update new stocksFII purchase
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some stocks purchase
...
  /stocks/sales/{id}:
    get:
      consumes:
      - application/json
      description: get stocks sale data
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Stock'
      summary: Get stocks sale by ID
    put:
      consumes:
      - application/json
      description: update new stocks sale
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some stocks sale
...
  /treasuries-direct/purchases/{id}:
    get:
      consumes:
      - application/json
      description: get treasury direct purchase data
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.TreasuryDirect'
      summary: Get treasury direct purchase by ID
    put:
      consumes:
      - application/json
      description: update new treasury direct purchase
      parameters:
      - description: Purchase id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some treasury direct purchase
...
  /treasuries-direct/sales/{id}:
    get:
      consumes:
      - application/json
      description: get treasury direct  sale data
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.TreasuryDirect'
      summary: Get treasury direct sale by ID
    put:
      consumes:
      - application/json
      description: update new treasury direct sale
      parameters:
      - description: Sale id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      summary: Update some treasury direct sale
swagger: "2.0"